# src/services/connector-service/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "_CreateConnectorRequest": {
        "additionalProperties": false,
        "description": "Body for ``POST /v1/workspaces/{ws}/connectors``.",
        "properties": {
          "credentialsAuthentication": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credentialsauthentication"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "leaseTtlSeconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leasettlseconds"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "targetConfig": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetconfig"
          },
          "type": {
            "minLength": 1,
            "title": "Type",
            "type": "string"
          },
          "usedCapabilities": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usedcapabilities"
          },
          "version": {
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "type",
          "version"
        ],
        "title": "_CreateConnectorRequest",
        "type": "object"
      },
...
      "_PatchConnectorRequest": {
        "additionalProperties": false,
        "description": "Body for ``PATCH /v1/workspaces/{ws}/connectors/{id}``.\n\nAll fields are optional; only present keys are forwarded to\n:meth:`InstanceService.patch`. Setting ``leaseTtlSeconds`` to\n``null`` clears the per-instance override and restores the\ndeployment default. The handler never silently drops unknown\nfields \u2014 Pydantic's ``extra=\"forbid\"`` rejects them at parse\ntime so the API surface speaks with one voice (rather than\ndelegating the \"unknown field\" 400 to ``ImmutableFieldUpdate``).",
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "leaseTtlSeconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leasettlseconds"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "_PatchConnectorRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Custos Connector Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/workspaces/{ws}/connector-types": {
      "get": {
        "description": "List every persisted version of a single connector type.\n\nConnector types themselves are global, but this route is mounted\nunder ``/v1/workspaces/{ws}/...`` so the operator surface has a\nconsistent shape and the call-context's ``workspace_id`` must\nstill match the ``{ws}`` path segment \u2014 that keeps the route\ntable workspace-pivoted even where the underlying data is not.\n\nReturns 200 with ``{items: [...], nextCursor: ...}``. The cursor\nis opaque; callers MUST pass it back verbatim to advance through\npages.",
        "operationId": "list_connector_type_versions_v1_workspaces__ws__connector_types_get",
        "parameters": [
          {
            "in": "path",
            "name": "ws",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Ws",
              "type": "string"
            }
          },
          {
            "description": "Connector type to list.",
            "in": "query",
            "name": "type",
            "required": true,
            "schema": {
              "description": "Connector type to list.",
              "minLength": 1,
              "title": "Type",
              "type": "string"
            }
          },
          {
            "description": "Opaque pagination cursor.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque pagination cursor.",
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 500,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List connector-type versions for a given type.",
        "tags": [
          "connector-types"
        ]
      }
    },
    "/v1/workspaces/{ws}/connectors": {
      "get": {
        "operationId": "list_connectors_v1_workspaces__ws__connectors_get",
        "parameters": [
          {
            "in": "path",
            "name": "ws",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Ws",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "enabled",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Enabled"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 500,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List connector instances in a workspace.",
        "tags": [
          "connectors"
        ]
      },
      "post": {
        "operationId": "create_connector_v1_workspaces__ws__connectors_post",
        "parameters": [
          {
            "in": "path",
            "name": "ws",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Ws",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_CreateConnectorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a connector instance.",
        "tags": [
          "connectors"
        ]
      }
    },
    "/v1/workspaces/{ws}/connectors/{instance_id}": {
      "get": {
        "operationId": "get_connector_v1_workspaces__ws__connectors__instance_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ws",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Ws",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Instance Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read a connector instance.",
        "tags": [
          "connectors"
        ]
      },
      "patch": {
        "operationId": "patch_connector_v1_workspaces__ws__connectors__instance_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "ws",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Ws",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Instance Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_PatchConnectorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Partially update a connector instance.",
        "tags": [
          "connectors"
        ]
      }
    },
...
  }
}